Take a stab at making pcb installations be relocatable.  By this I mean you can configure and install with
./configure --prefix=/original/path
make
make install
mv /original/path /new/path
and still have pcb work.

To make this work, the binary path is determined at runtime and all of the
other paths are computed relative to this.

As part of this, go ahead and kill off the pcb wrapper script around pcb-bin.
That was leftover from Xaw days and wasn't really needed anymore.  As a
side effect, the --program-prefix and --program-suffix configure options
will probably work now.