Add HID hook, close_confirm_dialog() for closing modified layouts.
This allows GUI HIDs to present a more native looking "save before closing"
dialog. The HID is responsible for any save which the user requests and
the subsequent return codes from the HID are as before, 0 for cancel and
1 for close.

This checkin also fixes the GTK HID's Save() action to return 1 for cancel
if the user cancels from the "Save As" dialog.