Don't bother unsubtracting complex polygons when doing incremental updates
Since the polygon is re-cleared against any objects existing within the
bounding box of the update region, there is no point wasting CPU cycles
intersecting a more complex polygon than necessary. Just take the UNION
of the polygon being unsubtracted, and the bounding box of the hole to
be filed in.