Add support for filling / thindrawing raw polygons to the HID interface
Refactor core polygon drawing to use these member functions, and split
the old code which used the HID's fill_polygon and draw_line functions
into a helper function which the guis now use to convert the polygon
into the primitives it knows how to draw.

Alter the NoHoles dicer to pass back PLINE contours, rather than
wrapping them in a POLYAREA and PolygonType.