Fix "newfullpoly" flag test and save its state into the .pcb file.
The flag controlling this behaviour is kept up to date as a PCB flag,
not something which is updated in Settings.FullPoly. Change the test
accordingly.

Added a PCB flag "newfullpoly" to save this state in the .pcb file.

NOTE: "full" polygons severly break connectivity checking, as the code
      always treats broken up pieces of the polygon as being connected.

      It _might_ have been better to leave this support broken - so
      users don't inadvertently create polygons with the "fullpoly" flag.

TODO: Consider removing, hiding or adding warnings to this feature.