From c5b84a8d1068baedebce85b8a4a9c39d4dd925e2 Mon Sep 17 00:00:00 2001 Subject: [PATCH] Fix a bug in gtk hid which caused layer buttons to get confused (sourceforge bug 1988951)
In ghid_layer_enable_buttons_update, the layer_buttons array was being
indexed using a counter which stoped at the current max_layer, which
was wrong.  Changed this to use the constants LAYER_BUTTON_*.

Also made the handling of the silk and rats buttons match the
(slightly more correct) handling of the other buttons in
layer_enable_button_cb.