draw.c: Draw all vias under the mask layer
This fixes an issue where the rendering of a partially masked via is
inaccurate. The down-side is that partially, or completely masked vias
become obscured by the mask drawn above them.

If the user is previewing the mask rendering, it is probably best that
they see an accurate representation - even if this means hiding vias.