hid/gtk: Prevent extra windows from showing unnecessarily
Prevent the library, log, and netlist windows in the GTK HID from
being shown unless explicitely requested.  This is done by splitting
ghid_*_window_show into ghid_*_window_create, which creates all the
resources for the window, and ghid_*_window_show, which actually shows
the window (and optionally "presents" it, which pulls it to the
foreground and gives it focus).

Since the NetlistShow action no longer actually shows the netlist
window, but sets the node/net that is shown in the window (if and when
it is visible), this patch also adds NetlistPresent which presents the
netlist window.  This can be added to the FindConnections menu item to
restore the behavior of bringing up the netlist window on a find.

Added an action LogShowOnAppend() which can restore the old behavior
that the log window is shown whenever something is appended to it
(currently off by default).