Remove Settings.init_done variable
This variable was previously used to allow action scripts run prior to
loading the GUI to avoid GUI startup after they call Quit(). The exact
semantics seem to date back prior to the HID split, where the Quit()
action would call gtk_main_quit(), an operation which is not legal if
the main loop was never started.

During the HID split, the Quit() action was changed to call exit(0),
so this dance isn't really necessary at all.