Speed up poly_ContourInContour() test by computing interior point
NB: This introduces a behaviour change in the boundary case, that two
identical contours will now be considered to be inside each other.

First perform a test on an arbitrary boundary node (proving that the
contour being testing for "insideness" is not outside the other
contour. (This cannot not conclusively prove the contour is inside).

In many cases, this simple node test gives enough evidence to return 0
for the ContourInContour test computing and testing an interior point.

Rather than checking each exterior point, compute a strictly interior
point (not on the boundary), and test that against the second contour.

Benchmarked to improve performance over other fixes for the buggy test.
Example board load (CPU) times for a complex board:

  21.50 (buggy contour_in_contour - single node point test)
  24.43 (brute-force node point tests)
  21.79 (single node test, then internal point test)