Fix poly_ComputeInteriorPoint() to work correctly for holes
The step where the algorithm finds a convex node to start from must
take into account whether the polygon vertices are ordered as a hole
or an outer contour. We now correctly compute a point inside the hole,
rather than possibly outside it.

This fixes an assertion on the following test-case. Prior to this
commit, the incorrect "interior" point tested for the concave hole
happens to lie inside the polygon's other hole, causing it to fail
an assert when processing the concave hole.

Layer(2 "solder")
(
  Line[340000 160000 183700 108000 1500 3000 "clearline"]
  Line[92000 121000 120000 90000 1500 3000 "clearline"]
  Line[270000 90000 120000 90000 1500 3000 "clearline"]
  Polygon("clearpoly")
  (
    [40000 40000] [320000 40000] [320000 200000] [40000 200000]
  )
)