Convert argument in HID_Flags to (void *) rather than (int)
void * allows us to pass pointers on all platforms, 64bit or
otherwise. We can still use casting macros to safely pass integer
values via this pointer.

Avoids the ugliness of castnig a size_t sized offset_of value into
an int. Due to the size of our structures, this did not cause any
actual bugs, but was not good practice.