flags.c: Avoid compiler warning when casting our OffsetOf value
Strictly, our OffsetOf calculation "could" return a number as large as the
pointer size on the system - so the compiler will complain when we squash
it directly into an int. Casting via a size_t will avoid the warning.

We know our structure offsets are small, so this is safe - if somewhat ugly.