find.c: Remove "undo" parameter to RatFindHook
Since all callers pass false for the "undo" parameter since the
last commit, remove it from the function prototype.

NB: The file-global variable "User", which it previously set can be
assumed to be false, as it defaults to false - and is restored after
any code which temporarily sets it to true.