Fix mesa crash on startup (for some mesa versions)
I'm not sure why it crashed, but it seems that providing a realize
handler which fiddles with the GL context solves it. I've also
removed a number of superfluous gtk_widget_realize() and an un-warraned
gtk_widget_show_all()

The underlying problem appears to have been fixed in later mesa versions.