Don't abuse CrosshairOn() API to force a redraw.
Expose DrawAttached() to the HIDs so they don't have to cheat.

Make DrawAttached() and DrawMark() a NOP if the relevant item is not
being shown. Ie. if !Crosshair.On, both functions draw nothing. If
!Marked.status, DrawMark() draws nothing.

A minor change is required in CrosshairOff() to XOR un-draw the item
before switching the flag to off.