hid/gtk: Drop in GL renderer - WOOT!
Much of the generic GL drawing stuff lives in a new common helper,
hid/common/hidgl.[ch]. Unavoidably, there is a lot of GUI specific
setup and teardown code.

There are probably still bits of code in hid/gtk/gtkhid-gl.c which
could be moved to a shared place if / when other HIDs wish to use
GL rendering.

Credits:

  Algorithm to calculate number of segments to use in circular
  curve approximation suggested by DJ Delorie.

  Thanks to Krzysztof Kościuszkiewicz for testing and debugging
  some issues with the GL_SCISSOR_TEST being used.