hid/gtk: Remove dead code (due to bitrot) from pinout_zoom_fit() function
The code tried to compare against the uninitialised, unused Output.Width
and Output.Height. Assuming these were initialised to zero by the compiler,
the conditional would always cause the function to return FALSE.

Remove the conditional, and drop the boolean return value from this
function, short-cutting the caller to the utilised code-path.

This removes the (long broken) feature where by the natural size
(in pixels) of the pinout preview would be reduced if it was greater than
3/4 of the viewport size (in pixels) of the active PCB.